Patches for Change Request Pass 5
