+++ src/multiselect.js
@@ -10,11 +10,9 @@
 
 import * as Blockly from 'blockly/core';
 
-import DragSelect from '../lib/ds.min';
-
 import * as ContextMenu from './multiselect_contextmenu';
 import * as Shortcut from './multiselect_shortcut';
-import {blockSelection, inMultipleSelectionMode, setSelectionMode,
+import {blockSelection, inMultipleSelectionMode,
   hasSelectedParent} from './global';
 import {MultiselectControls} from './multiselect_controls';
 
@@ -28,8 +26,6 @@ export class Multiselect {
    */
   constructor(workspace) {
     this.workspace_ = workspace;
-    this.hasDisableWorkspaceDrag_ = false;
-    this.justUnselectedBlock_ = null;
     this.origHandleWsStart_ = Blockly.Gesture.prototype.handleWsStart;
   }
 
@@ -40,9 +36,9 @@ export class Multiselect {
    */
   init(options) {
     this.onKeyDownWrapper_ = Blockly.browserEvents.conditionalBind(
-        this.workspace_.injectionDiv_, 'keydown', this, this.onKeyDown_);
+        this.workspace_.getInjectionDiv(), 'keydown', this, this.onKeyDown_);
     this.onKeyUpWrapper_ = Blockly.browserEvents.conditionalBind(
-        this.workspace_.injectionDiv_, 'keyup', this, this.onKeyUp_);
+        this.workspace_.getInjectionDiv(), 'keyup', this, this.onKeyUp_);
     this.eventListenerWrapper_ = this.eventListener_.bind(this);
     this.workspace_.addChangeListener(this.eventListenerWrapper_);
 
@@ -51,13 +47,11 @@ export class Multiselect {
     Shortcut.unregisterShortcut();
     Shortcut.registerOurShortcut();
 
-    if (options.multiselectIcon && options.multiselectIcon.enabled) {
-      this.controls_ = new MultiselectControls(
-          this.workspace_, options.multiselectIcon, this);
-      const svgControls = this.controls_.createDom();
-      this.workspace_.svgGroup_.appendChild(svgControls);
-      this.controls_.init();
-    }
+    this.controls_ = new MultiselectControls(
+        this.workspace_, options.multiselectIcon, this);
+    const svgControls = this.controls_.createDom();
+    this.workspace_.getParentSvg().appendChild(svgControls);
+    this.controls_.init();
 
     if (options.useDoubleClick) {
       this.useDoubleClick_(true);
@@ -168,73 +162,15 @@ export class Multiselect {
   }
 
   /**
-   * Maintain the selected blocks set list when updating.
-   * @param {!Blockly.BlockSvg} block The block to update.
-   * @private
-   */
-  updateBlocks_(block) {
-    if (block &&
-      block.isDeletable() &&
-      block.isMovable()) {
-      if (blockSelection.has(block.id)) {
-        blockSelection.delete(block.id);
-        this.justUnselectedBlock_ = block;
-        block.pathObject.updateSelected(false);
-      } else {
-        blockSelection.add(block.id);
-        this.justUnselectedBlock_ = null;
-        block.pathObject.updateSelected(true);
-        block.bringToFront();
-      }
-    }
-  }
-
-  /**
    * Handle workspace events.
    * @param {!Event} e Blockly event.
    * @private
    */
   eventListener_(e) {
     // on Block Selected
-    if (e.type !== Blockly.Events.SELECTED) {
-      return;
+    if (e.type === Blockly.Events.SELECTED) {
+      this.controls_.updateMultiselect();
     }
-
-    if (!inMultipleSelectionMode) {
-      if (!Blockly.selected ||
-        (Blockly.selected && !blockSelection.has(Blockly.selected.id))) {
-        // When not in multiple selection mode and Blockly selects a block not
-        // in currently selected set or unselects, clear the selected set.
-        blockSelection.forEach((id) => {
-          const element = this.workspace_.getBlockById(id);
-          if (element) {
-            element.pathObject.updateSelected(false);
-          }
-        });
-        blockSelection.clear();
-        this.updateBlocks_(Blockly.selected);
-      }
-    } else if (this.justUnselectedBlock_ && Blockly.selected &&
-      Blockly.selected.id === this.justUnselectedBlock_.id) {
-      // Update the Blockly selected block when that block is
-      // no longer selected in our set.
-      if (blockSelection.size) {
-        Blockly.common.setSelected(
-            this.workspace_.getBlockById(blockSelection.keys().next().value));
-      } else {
-        Blockly.common.setSelected(null);
-      }
-      this.justUnselectedBlock_.pathObject.updateSelected(false);
-      this.justUnselectedBlock_ = null;
-    }
-
-    // Update the selection highlight.
-    blockSelection.forEach((id) => {
-      const element = this.workspace_.getBlockById(id);
-      if (element) {
-        element.pathObject.updateSelected(true);
-      }
-    });
   }
 
   /**
@@ -243,48 +179,10 @@ export class Multiselect {
    * @private
    */
   onKeyDown_(e) {
-    if (e.keyCode !== Blockly.utils.KeyCodes.SHIFT ||
-        inMultipleSelectionMode) {
-      return;
+    if (e.keyCode === Blockly.utils.KeyCodes.SHIFT &&
+        !inMultipleSelectionMode) {
+      this.controls_.enableMultiselect();
     }
-
-    // Ensure that we only restore drag to move the workspace behavior
-    // when it is enabled.
-    if (!this.hasDisableWorkspaceDrag_ &&
-      this.workspace_.options.moveOptions &&
-      this.workspace_.options.moveOptions.drag) {
-      this.workspace_.options.moveOptions.drag = false;
-      this.hasDisableWorkspaceDrag_ = true;
-    }
-    this.dragSelect_ = new DragSelect({
-      selectables: document.querySelectorAll(
-          'g.blocklyDraggable:not(.blocklyInsertionMarker)' +
-        '> path.blocklyPath'),
-      area: document.querySelector('.blocklyWorkspace'),
-      multiselectMode: true,
-      draggability: false,
-      usePointerEvents: true,
-    });
-    this.dragSelect_.subscribe('elementselect', (info) => {
-      const element = info.item.parentElement;
-      if (inMultipleSelectionMode && element.dataset &&
-        element.dataset.id) {
-        this.updateBlocks_(
-            this.workspace_.getBlockById(element.dataset.id));
-      }
-    });
-    this.dragSelect_.subscribe('elementunselect', (info) => {
-      const element = info.item.parentElement;
-      if (inMultipleSelectionMode && element.dataset &&
-        element.dataset.id) {
-        this.updateBlocks_(
-            this.workspace_.getBlockById(element.dataset.id));
-      }
-    });
-    if (this.controls_) {
-      this.controls_.updateMultiselectIcon(true);
-    }
-    setSelectionMode(true);
   }
 
   /**
@@ -293,29 +191,8 @@ export class Multiselect {
    * @private
    */
   onKeyUp_(e) {
-    if (e.keyCode !== Blockly.utils.KeyCodes.SHIFT) {
-      return;
-    }
-
-    setSelectionMode(false);
-    if (this.dragSelect_) {
-      this.dragSelect_.stop();
-      this.dragSelect_ = null;
-    }
-    // Ensure that at least Blockly select one of the blocks in the
-    // selection set, or clear the Blockly selection if our set is empty.
-    if (!blockSelection.size && Blockly.selected) {
-      Blockly.common.setSelected(null);
-    } else if (blockSelection.size && !Blockly.selected) {
-      Blockly.common.setSelected(
-          this.workspace_.getBlockById(blockSelection.keys().next().value));
-    }
-    if (this.hasDisableWorkspaceDrag_) {
-      this.workspace_.options.moveOptions.drag = true;
-      this.hasDisableWorkspaceDrag_ = false;
-    }
-    if (this.controls_) {
-      this.controls_.updateMultiselectIcon(false);
+    if (e.keyCode === Blockly.utils.KeyCodes.SHIFT) {
+      this.controls_.disableMultiselect();
     }
   }
 }
